finish first stage of basic springboot app (#2)
* feat: 🎸 added DTO for return of Employee

- added EmployeeDTO for return without ID, createDate -> not need to
change when POSTing

* feat: 🎸 Added Paging for getting Employees

- add Paging

* feat: 🎸 added prod-env to direct use VPC on GCP

finish GCP RUN deployement by using VPC, added auto-forward function in
yml file(auto trans to https), added stuff in Dockerfile to make it
compile itself not pre-compile then build.

+++ src/main/java/tw/com/cathaybk/service/EmployeeService.java
@@ -1,51 +1,96 @@
 package tw.com.cathaybk.service;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import tw.com.cathaybk.dto.EmployeeDTO;
 import tw.com.cathaybk.entity.Employee;
 import tw.com.cathaybk.dao.EmployeeDAO;
 import tw.com.cathaybk.exception.UserNotFoundException;
 import tw.com.cathaybk.utils.Result;
 
-import java.util.ArrayList;
-import java.util.List;
+import java.time.LocalDateTime;
+
 
 @Service
+@Slf4j
 public class EmployeeService {
 
     @Autowired
     private EmployeeDAO employeeDAO;
 
-    public Result<Employee> saveEmployee(Employee employee) {
-        employeeDAO.save(employee); //TODO 因為新增也是自增，不會需要帶入employeeID -> 可能還是拿DTO好
-        return Result.success(employee);
+    /**
+     * 儲存資料，回傳員工資訊
+     * @param employee
+     * @return Result
+     */
+    public Result<Employee> saveEmployee(EmployeeDTO employee) {
+        Employee emp = new Employee();
+        BeanUtils.copyProperties(employee, emp);
+        emp.setCreateTime(LocalDateTime.now());
+        emp.setUpdateTime(LocalDateTime.now());
+        employeeDAO.save(emp);
+        log.info("員工添加成功 id: {}", emp.getEmployeeId());
+        return Result.success(emp);
     }
 
-    public Result<List<Employee>> getEmployees() {
-        List<Employee> employees = new ArrayList<>();
-        employeeDAO.findAll().forEach(employees::add);
-        return Result.success(employees);
+    /**
+     * 取得所有員工資料，會自動分頁
+     * @param pageable
+     * @return Result of Pages
+     */
+    public Result<Page<Employee>> getEmployees(Pageable pageable) {
+        return Result.success(employeeDAO.findAll(pageable));
     }
 
+    /**
+     * 一個通用method，找到則回傳員工資料，失敗進入ExceptionHandler
+     * @param id
+     * @return Employee
+     */
     private Employee getEmployeeById(Long id) {
+        log.info("取得員工資料 id: {}", id);
         return employeeDAO.findById(id).orElseThrow(() -> new UserNotFoundException("User Not Found"));
     }
 
+    /**
+     * 根據id取得員工資料
+     * @param id
+     * @return Result
+     */
     public Result<Employee> getEmployee(Long id) {
         return Result.success(getEmployeeById(id));
     }
 
-    public Result<Employee> deleteEmployee(Long id) {
-        Employee employee = getEmployeeById(id);
+    /**
+     * 刪除員工
+     * @param id
+     * @return Result
+     */
+    public Result deleteEmployee(Long id) {
+        if(!employeeDAO.existsById(id)) {
+            log.info("查無員工資料 id: {}", id);
+            throw new UserNotFoundException("User Not Found");
+        }
+        log.info("員工資料刪除成功 id: {}", id);
         employeeDAO.deleteById(id);
-        return Result.success(employee);
+        return Result.success();
     }
 
-    public Result<Employee> updateEmployee(Long id, Employee employee) {
+    /**
+     * 更改員工資料
+     * @param id
+     * @param employee
+     * @return Result，如果失敗則會進入Result.error(Exception)
+     */
+    public Result<Employee> updateEmployee(Long id, EmployeeDTO employee) {
         Employee emp = getEmployeeById(id);
         BeanUtils.copyProperties(employee, emp);
-        emp.setEmployeeId(id); //TODO 考慮用DTO取代掉這一段，便可以不用在封裝時傳入id
+        emp.setUpdateTime(LocalDateTime.now());
+        log.info("員工資料更新成功 id: {}", emp.getEmployeeId());
         return Result.success(employeeDAO.save(emp)); //try to use return employeeDAO.save(employee) but fail -> PID duplicate
     }
 }
